pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eaip-viewer"
version = "0.1.0"
description = "Electronic Aeronautical Information Publication Viewer"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Cg8-5712", email = "5712.cg8@gmail.com"},
    {name = "HTony03", email = "HTony03@foxmail.com"}
]
keywords = ["aviation", "eaip", "charts", "viewer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PyQt6>=6.4.0",
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "cryptography>=41.0.0",
    "SQLAlchemy>=2.0.0",
    "alembic>=1.12.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "platformdirs>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pyinstaller>=5.13.0",
]

[project.urls]
Homepage = "https://github.com/cg8-5712/EAIP_Viewer"
Repository = "https://github.com/cg8-5712/EAIP_Viewer"
Documentation = "https://github.com/cg8-5712/EAIP_Viewer/tree/main/docs"
Issues = "https://github.com/cg8-5712/EAIP_Viewer/issues"

[project.scripts]
eaip-viewer = "src.main:main"

# ==================== Black 配置 ====================
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | \.pytest_cache
  | build
  | dist
)/
'''

# ==================== isort 配置 ====================
[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

# ==================== mypy 配置 ====================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = [
    "tests/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = [
    "PyQt6.*",
    "fitz.*",
]
ignore_missing_imports = true

# ==================== pytest 配置 ====================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

# ==================== coverage 配置 ====================
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
